NEW Initial CLI with the start of the AST visiting system.
First clone then analyse. Allows for making changes to
the cloned stuff if necessary but more importantly to analyse repeatedly
over the same codebases (potentially with varying settings).

Idea is parse both versions of the same file, then look at the
differences.
At some stage I'll need a "check the superclass" step - Unsure yet
whether it'll be "go hunt for that superclass and check it right now"
during the initial analysis or if it'll be a separate "Go through all
the collected class-like data and check superclasses" after the fact.